api/requirements.txt
fastapi>=0.109.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
uvicorn>=0.27.0
requests-oauthlib>=1.3.1
email-validator>=2.0.0
aiosmtplib>=3.0.0